Added file option
Added a dropdown to select different files for viewing, will load file on select and store it in a dictionary so that each file selected is only loaded once. Also refactored layout to not use a table, axis selection would have used flex-grow for allocating width space but seems to be some issue with attribute not appearing in div when rendered so set widths as %30 each.